requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
entsoe-py>=0.6.0
openai>=1.0.0
matplotlib>=3.7.0
shap>=0.43.0
python-dotenv>=1.0.0
